[web] Refactor webparagraph painters to separate CK properly
