+++ engine/src/flutter/lib/web_ui/lib/src/engine.dart
@@ -40,6 +40,7 @@ export 'engine/canvaskit/surface.dart';
 export 'engine/canvaskit/text.dart';
 export 'engine/canvaskit/util.dart';
 export 'engine/canvaskit/vertices.dart';
+export 'engine/canvaskit/web_paragraph_painter.dart';
 export 'engine/clipboard.dart';
 export 'engine/color_filter.dart';
 export 'engine/compositing/canvas_provider.dart';
@@ -168,10 +169,9 @@ export 'engine/view_embedder/style_manager.dart';
 export 'engine/web_paragraph/bidi.dart';
 export 'engine/web_paragraph/code_unit_flags.dart';
 export 'engine/web_paragraph/debug.dart';
+export 'engine/web_paragraph/decorations.dart';
 export 'engine/web_paragraph/font_collection.dart';
 export 'engine/web_paragraph/layout.dart';
-export 'engine/web_paragraph/paint.dart';
-export 'engine/web_paragraph/paint_paragraph.dart';
 export 'engine/web_paragraph/painter.dart';
 export 'engine/web_paragraph/paragraph.dart';
 export 'engine/web_paragraph/wrapper.dart';
